The rt package

rt provides a programming interface to the Request Tracker API.

Details

Everything should be implemented and all functions should return a reasonably useful result that's suitable for integrating into your workflows.

Setup

Before you can do anything useful with this package, you'll need to do three things:

Determine your base URL and set it using Sys.setenv(RT_BASE_URL="your url here). In most cases, this will be the same as the URL of the page you use to log in to RT.

Determine and set your credentials. You can skip setting them if you like and skip to step 3 or you can set them using the RT_USER and RT_PASSWORD environmental variables via Sys.setenv. See rt_login for more.

Log in using R by calling rt_login. See rt_login for more.

A typical flow for setting up your R session to work with RT might look like this:

Sys.setenv(RT_BASE_URL = "http://example.com/rt",
           RT_USER = "me@example.com",
           RT_PASSWORD = "mypassword")

If you use RT a lot, you might consider putting code like the above in your .Renviron, minus the call to rt_login() so the environmental variables are available but you aren't logging into RT every time you start R. See ?Startup for more information.

Available Functions

General

rt_login

rt_logout

Tickets

rt_ticket_search

rt_ticket_create

rt_ticket_edit

rt_ticket_history

rt_ticket_history_comment

rt_ticket_history_reply

rt_ticket_links

rt_ticket_links_edit

rt_ticket_merge

rt_ticket_properties

rt_ticket_attachments

rt_ticket_attachment

rt_ticket_attachment_content

Users

rt_user_create

rt_user_edit

rt_user_properties

Queues

rt_queue_properties